Update index.js, update readme and demo.gif
