+++ example/App.js
@@ -8,7 +8,7 @@ import {
     Dimensions,
 } from 'react-native';
 
-import ImageView from './ImageCollectionView';
+import ImageView from './ImageView';
 // import ImageView from 'react-native-image-view';
 
 const {width} = Dimensions.get('window');
